Introduce stub on view model
